modularize into Chart component
